requirements.txt
django-appconf>=1.0
django-countries
django-vies
fiobank
python-dateutil
requests
thepay
